+++ tests/Handler/ProjectListHandlerTest.php
@@ -19,7 +19,8 @@ class ProjectListHandlerTest extends CommandTestCase
         parent::setUp();
 
         TestKernel::$jiraService = $this->jiraService;
-        $this->handler = new ProjectListHandler($this->jiraService);
+        TestKernel::$translationService = $this->translationService;
+        $this->handler = new ProjectListHandler($this->jiraService, $this->translationService);
     }
 
     public function testHandle(): void